requirements.txt
python-dotenv~=1.0.1
requests~=2.32.3
simple-salesforce~=1.12.6
pandas~=2.2.3
click~=8.1.7